duplicate tagName function
