+++ src/renderers/react/react.ts
@@ -1,12 +1,27 @@
-import { tagName } from './shared';
 import Tag from '../../tag';
 import { RenderableTreeNodes, Scalar } from '../../types';
-import type { createElement, Fragment, ReactNode } from 'react';
+
+import type { createElement, ComponentType, Fragment, ReactNode } from 'react';
 
 type ReactShape = Readonly<{
   createElement: typeof createElement;
   Fragment: typeof Fragment;
 }>;
+
+type Component = ComponentType<unknown>;
+
+function tagName(
+  name: string,
+  components: Record<string, Component> | ((string: string) => Component)
+): string | Component {
+  return typeof name !== 'string'
+    ? name // This can be an object, e.g. when React.forwardRef is used
+    : name[0] !== name[0].toUpperCase()
+    ? name
+    : components instanceof Function
+    ? components(name)
+    : components[name];
+}
 
 export default function dynamic(
   node: RenderableTreeNodes,
